incode/app/requirements.txt
flask
mysql-connector-python
redis-py-cluster
flask-session